pyproject.toml
[project]
name = "mcp-serper-server"
version = "0.1.0"
description = "mcp server with serper intigration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.10.1",
    "pdfplumber>=0.11.7",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "qrcode[pil]>=8.2",
    "requests>=2.32.4",
]
